refractory  the create transaction request scala tests
